requirements.txt
attrs==21.4.0
django==5.1.1
djangorestframework==3.14.0
psycopg2>=2.9.6
cached-property==1.5.2
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
isodate==0.6.1
lxml==4.8.0
platformdirs==2.5.1
pytz==2021.3
requests==2.27.1
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.16.0
urllib3==1.26.8
